chore(invariant): config defaults, add max test threads config (#7957)
* chore(invariant): default depth / shrink run limit, config num of threads

* Limit depth in testing at 15

* Set depth for ext integrations to 15

* Make max threads global config, propagate error if thread pool creation fails

* Run invariant gas report with depth of 15

* Typo, move set threads at top of execute

+++ crates/evm/evm/src/executors/invariant/error.rs
@@ -64,7 +64,7 @@ pub struct FailedInvariantCaseData {
     /// Inner fuzzing Sequence coming from overriding calls.
     pub inner_sequence: Vec<Option<BasicTxDetails>>,
     /// Shrink run limit
-    pub shrink_run_limit: usize,
+    pub shrink_run_limit: u32,
     /// Fail on revert, used to check sequence when shrinking.
     pub fail_on_revert: bool,
 }
